# Separator improvements