* separator improvements
    * add support for [Windows/Mac OS 9 &c. line endings](https://en.wikipedia.org/wiki/Newline#Representations)
    * allow the line separator to be specified
    * use the (slightly faster) old append method if the line separator is a string
* ignore null/undefined
* comprehensive tests
* add benchmark
* README.md: update package name: `outputbuffer` -> `output-buffer`