@achingbrain 2015-02-09
Thanks for sending this through.  Allowing people to specify the separator seems like a good idea, and thanks for the unit tests.  I'm quite keen to get this merged.

A few questions though:

1. Could you explain the benchmark suite a little more - it seems to compare the new string/regex split functions with the current implementation.  Once this request is merged it'll alter the current implementation so having it hanging around in a benchmark seems redundant.  Also the coding style is inconsistent with the rest of the module (not a big problem, just a little jarring).

2. Running the benchmarks a few times doesn't seem to show a whole lot of difference between the current and proposed implementations:

  ```
  string-split buffer x 159,269 ops/sec ±0.84% (87 runs sampled)
  regex-split buffer  x 130,827 ops/sec ±1.81% (74 runs sampled)
  old buffer          x 157,780 ops/sec ±0.84% (85 runs sampled)

  string-split buffer x 148,329 ops/sec ±4.71% (81 runs sampled)
  regex-split buffer  x 128,725 ops/sec ±2.03% (89 runs sampled)
  old buffer          x 159,370 ops/sec ±1.65% (66 runs sampled)

  string-split buffer x 159,017 ops/sec ±1.50% (80 runs sampled)
  regex-split buffer  x 133,441 ops/sec ±1.12% (88 runs sampled)
  old buffer          x 159,509 ops/sec ±1.98% (67 runs sampled)

  string-split buffer x 160,199 ops/sec ±0.90% (86 runs sampled)
  regex-split buffer  x 134,102 ops/sec ±1.01% (90 runs sampled)
  old buffer          x 159,221 ops/sec ±0.91% (86 runs sampled)
  ```
  Have you observed something different?

3. It looks like you've tried to optimise the code (declaring variables up front, storing array lengths, using switch instead of if/else etc):

  ```javascript
  function regexSplitAppend(data) {
    var lines = data.split(this._sep)
    var len = lines.length
    var last, i

    switch (len) {
        ...
      default
        last = len - 1
        ...
        for (i = 1; i < last; ++i) {
          ...
        }
        this._buffer = lines[last]
        ...
    }
  }
  ```

  vs

  ```javascript
  function regexSplitAppend(data) {
    var lines = data.split(this._sep)

    if(lines.length == 0) {
      ...
    } else {
      var last = lines.length - 1
      ...
      for(var i = 0; i < last; i++) {
         ...
      }
      this._buffer = lines[last]
      ...
    }
  }
  ```
  To my eyes the second is more readable because the variables are declared close to where they will be used (and you'd hope that since v8 knows then arrays are modified, it has some sensible array length caching and is not counting every array entry every time `lines.length` is accessed), however I'm willing to sacrifice what I perceive as readability if it solves a problem.

  I ran the benchmark suite with both versions above and again the results were not wildly different.  Have you been seeing performance problems?