@chocolateboy 2015-02-09
> Could you explain the benchmark suite a little more

It's just for information. I would have preferred to use just the split/switch version since I find it easier to read/understand, but I see the virtue of going that extra optimization mile in library code, so I kept both. There's no need to keep the benchmark in the repo.

> Have you observed something different?

No.

> To my eyes the second is more readable... however I'm willing to sacrifice what I perceive as readability if it solves a problem.

I don't mind either way.

> Have you been seeing performance problems?

No.