@achingbrain 2015-02-09
Great stuff, thanks.  One last thing - [here](https://github.com/achingbrain/output-buffer/pull/1/files#diff-6be479e7aa090e765e311e28a4c1f775R46):

```javascript
this._append = ({}.toString.call(this._sep) === '[object RegExp]') ? regexSplitAppend : stringSplitAppend
```

Is there any reason you didn't just use `instanceof`?  E.g.:

```javascript
this._append = this._sep instanceof RegExp ? regexSplitAppend : stringSplitAppend
```