@chocolateboy 2015-02-09
A couple of results from my system:

```
  string-split buffer x 62,938 ops/sec ±5.77% (69 runs sampled)
  regex-split buffer  x 57,070 ops/sec ±2.20% (85 runs sampled)
  old buffer          x 65,269 ops/sec ±2.00% (80 runs sampled)
```

```
  string-split buffer x 62,116 ops/sec ±3.29% (68 runs sampled)
  regex-split buffer  x 55,219 ops/sec ±2.28% (81 runs sampled)
  old buffer          x 64,044 ops/sec ±2.53% (79 runs sampled)
```

```
  string-split buffer x 63,954 ops/sec ±3.23% (69 runs sampled)
  regex-split buffer  x 56,096 ops/sec ±2.04% (83 runs sampled)
  old buffer          x 64,806 ops/sec ±2.07% (80 runs sampled)
```