@chocolateboy 2015-02-09
> Is there any reason you didn't just use `instanceof`?

I can't think of one, but the [lodash code for `isRegExp`](https://github.com/lodash/lodash/blob/3.1.0/lodash.src.js#L8238) suggests there might be one (perhaps performance).