@achingbrain 2015-02-09
Thinking about it, it might be better to just pass `this._sep` to [String.split](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/String/split) and have done with it since it takes a String or a RegExp.

```javascript
'foo\nbar\nbaz'.split('\n')
> ["foo", "bar", "baz"]
```

```javascript
'foo\nbar\nbaz'.split(/\n/)
> ["foo", "bar", "baz"]
```

```javascript
'foo\nbar\nbaz'.split('/\n/')
> ["foo", "bar", "baz"]
```