@chocolateboy 2015-02-09
That was the original design, but it's still (slightly) slower than the dedicated string-split code. Up to you :-)